add macd& update backtest

+++ service/include/DataSource.h
@@ -10,7 +10,7 @@ struct DataFeatures {
     Vector<size_t> _features;
     symbol_t _symbol;
     double _price;
-    Vector<float> _data;
+    Vector<double> _data;
     Vector<int64_t> _dimesion;
     YAS_DEFINE_STRUCT_SERIALIZE("DataMessenger", _symbol, _features, _price, _data, _dimesion);
 };
