+++ service/include/std_header.h
@@ -74,4 +74,4 @@ using UnorderedSet = std::unordered_set<T, Args...>;
 #include "Eigen/Core"
 #include "Util/log.h"
 
-using feature_t = std::variant<double, Vector<float>>;
+using feature_t = std::variant<double, Vector<double>>;
